+++ scripts/delayed.js
@@ -1,20 +1,5 @@
 // delay loading of GTM script until after the page has loaded
-import { isProd } from '../libs/utils/utils.js';
-
-const DEV_LAUNCH_SCRIPT = 'https://assets.adobedtm.com/ad9123205592/67641f4a9897/launch-b238893bfd09-staging.min.js';
-const PROD_LAUNCH_SCRIPT = 'https://assets.adobedtm.com/ad9123205592/67641f4a9897/launch-fc986eef9273.min.js';
-
-function loadAdobeLaunch() {
-  const tag = document.createElement('script');
-  tag.type = 'text/javascript';
-  tag.async = true;
-  if (isProd()) {
-    tag.src = PROD_LAUNCH_SCRIPT;
-  } else {
-    tag.src = DEV_LAUNCH_SCRIPT;
-  }
-  document.querySelector('head').append(tag);
-}
+import { isProductionEnvironment } from '../libs/utils/utils.js';
 
 function loadGoogleTagManagerDev() {
   const gtmScript = document.createElement('script');
@@ -75,8 +60,7 @@ function loadFullStoryProd() {
 }
 
 if (window.location.hostname !== 'localhost') {
-  loadAdobeLaunch();
-  if (isProd()) {
+  if (isProductionEnvironment()) {
     loadGoogleTagManagerProd();
     loadFullStoryProd();
   } else {
