+++ scripts/scripts.js
@@ -14,7 +14,7 @@ import {
 } from './aem.js';
 
 import { decorateButtons } from '../libs/utils/decorate.js';
-import { loadPalette, createTag } from '../libs/utils/utils.js';
+import { loadPalette, createTag, isProductionEnvironment } from '../libs/utils/utils.js';
 
 export const PRODUCTION_DOMAINS = ['www.creditacceptance.com'];
 
@@ -424,6 +424,21 @@ async function loadEager(doc) {
   }
 }
 
+const DEV_LAUNCH_SCRIPT = 'https://assets.adobedtm.com/ad9123205592/67641f4a9897/launch-b238893bfd09-staging.min.js';
+const PROD_LAUNCH_SCRIPT = 'https://assets.adobedtm.com/ad9123205592/67641f4a9897/launch-fc986eef9273.min.js';
+
+function loadAdobeLaunch() {
+  const tag = document.createElement('script');
+  tag.type = 'text/javascript';
+  tag.async = true;
+  if (isProductionEnvironment()) {
+    tag.src = PROD_LAUNCH_SCRIPT;
+  } else {
+    tag.src = DEV_LAUNCH_SCRIPT;
+  }
+  document.querySelector('head').append(tag);
+}
+
 /**
  * Loads everything that doesn't need to be delayed.
  * @param {Element} doc The container element
@@ -443,6 +458,9 @@ async function loadLazy(doc) {
   await loadPalette();
   loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
   loadFonts();
+  if (window.location.hostname !== 'localhost') {
+    loadAdobeLaunch();
+  }
 }
 
 /**
